Move builder into new file and some fn typedefs into a separate file. Implement `Init::default` for `Mdl`s that `impl Default`
